refactor(suite-native): extract backup selectors into @suite-native/backup package
