+++ suite-common/device/src/deviceSelectors.ts
@@ -42,7 +42,6 @@ import { versionUtils } from '@trezor/utils';
 
 import {
     DEVICE_LOW_BATTERY_PERCENTAGE_THRESHOLD,
-    EXTENDABLE_SHAMIR_BACKUP_TYPES,
     PORTFOLIO_TRACKER_DEVICE_ID,
 } from './deviceConstants';
 import { type DeviceRootState } from './deviceReducer';
@@ -152,22 +151,6 @@ export const selectIsDeviceBackupRequired = createMemoizedSelector(
 export const selectIsDeviceBackupUnfinished = createMemoizedSelector(
     [selectDeviceFeatures],
     features => features?.unfinished_backup === true,
-);
-
-export const selectIsAdditionalShamirBackupInProgress = createMemoizedSelector(
-    [selectDeviceFeatures],
-    features =>
-        features?.recovery_status === 'Backup' &&
-        features.recovery_type === undefined &&
-        features.backup_availability === 'Available',
-);
-
-export const selectHasExtendableShamirBackup = createMemoizedSelector(
-    [selectDeviceFeatures],
-    features =>
-        features?.backup_type !== undefined &&
-        features?.backup_type !== null &&
-        EXTENDABLE_SHAMIR_BACKUP_TYPES.includes(features.backup_type),
 );
 
 export const selectDeviceLanguage = createMemoizedSelector(
@@ -446,14 +429,6 @@ export const selectDeviceModelById = createMemoizedSelector(
 
 export const selectDeviceModel = createMemoizedSelector([selectSelectedDevice], selectedDevice =>
     selectedDevice ? getDeviceInternalModel(selectedDevice) : null,
-);
-
-export const selectIsCreateAdditionalBackupAvailable = createMemoizedSelector(
-    [selectDeviceModel, selectHasExtendableShamirBackup, selectDeviceFeatures],
-    (deviceModel, hasExtendableShamirBackup, features) =>
-        deviceModel === DeviceModelInternal.T3W1 && // NOTE: FW will expose a capability flag for this, so it should be replaced with that than
-        hasExtendableShamirBackup &&
-        features?.backup_availability === 'NotAvailable',
 );
 
 export const selectFirmwareReleaseConfig = createMemoizedSelector(
