+++ suite-native/backup/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [
+        { "path": "../../suite-common/device" },
+        {
+            "path": "../../suite-common/redux-utils"
+        },
+        { "path": "../../packages/connect" },
+        { "path": "../../packages/device-utils" }
+    ]
+}
