+++ suite-native/module-device-settings/package.json
@@ -24,6 +24,7 @@
         "@suite-common/wallet-core": "workspace:*",
         "@suite-native/alerts": "workspace:*",
         "@suite-native/atoms": "workspace:*",
+        "@suite-native/backup": "workspace:*",
         "@suite-native/bluetooth": "workspace:*",
         "@suite-native/device": "workspace:*",
         "@suite-native/device-authorization": "workspace:*",
