+++ suite-native/nfc/tsconfig.json
@@ -6,6 +6,7 @@
         {
             "path": "../../suite-common/redux-utils"
         },
+        { "path": "../backup" },
         { "path": "../atoms" },
         { "path": "../device-mutex" },
         { "path": "../intl" },
